requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
dohq-tfs==1.0.175
Flask==1.1.2
glob2==0.7
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.3.0
ntlm-auth==1.4.0
packaging==20.3
parse==1.14.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.2
pytest-bdd==3.2.1
requests==2.23.0
requests-ntlm==1.1.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
